requirements.txt
iota_sdk==1.1.4
pycryptodome==3.22.0
PyOTA==2.0.7
python-dotenv==1.1.0
Requests==2.32.3